misc/PrivacyInAction/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
requests>=2.31.0

# Azure authentication (choose based on your setup)
# Option 1: For service principal or managed identity
azure-identity>=1.15.0

# Option 2: For interactive browser authentication
# azure-identity[interactive]>=1.15.0

# Note: The 'agents' and 'mcp' packages need to be verified:
# - 'agents' appears to be Microsoft Agents SDK - check actual package name
# - 'mcp' is Model Context Protocol SDK - verify installation method
# These may need to be installed from specific sources or have different names

